requirements.txt
ijson==3.2.3
matplotlib==3.7.5
numpy==1.24.4
pandas==1.5.3
pillow==9.5.0
plotly==5.19.0
pycocotools==2.0.7
seaborn==0.13.2
streamlit==1.31.1
typing-extensions==4.7.1
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.0+cu118
torchvision==0.15.0+cu118
tqdm==4.66.1
